Create Link to download
I was working on a recommendation system that depended on the Surprise library. However, since I was using the latest Python version (3.11), I encountered compatibility issues. If you are using Python 3.6, you can directly install the library using pip. If you are unable to do that, you can use this pre-compiled module instead.